improcve chart formatting
